Create speech-to-text_geocoder
The process of making this project first involved importing SpeechRecognition, PyAudio modules, and geocoder. Then after we had initialized the Speech Recognition module, we created a Microphone object which would record the speech of the said person, dictating the location. Then, the speech is converted to text and the final destination is extracted from it. Following this, there is a route generated from the current location of the individual to the desired final destination geocoding and Google Maps.